# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Parents API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://3.132.243.209/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Categories",
                "operationId": "Categories",
                "responses": {
                    "200": {
                        "description": "Categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriesResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Products",
                "operationId": "Products",
                "responses": {
                    "200": {
                        "description": "Products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BadRequest": {
                "title": "Bad Request",
                "description": "BadRequest",
                "properties": {
                    "dsfsfds": {
                        "title": "dsfsfds",
                        "description": "Bad Request",
                        "type": "string"
                    }
                },
                "type": "string"
            },
...
            "ProductResponse": {
                "title": "Categories Response",
                "description": "Categories Response",
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "id": "2",
                            "name": "test test",
                            "slug": "test-test",
                            "status": "1",
                            "weight": "21.00",
                            "description": "dfsdfsdfsd",
                            "sku": "dsfd",
                            "quantity": "12",
                            "price": "12.00",
                            "sale_price": "121.00",
                            "variation_attributes": "",
                            "created": "2021-12-13",
                            "updated_at": "2021-12-13"
                        }
                    }
                },
                "type": "object"
            },
...
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden",
                "properties": {
                    "forbidden": {
                        "title": "forbidden",
                        "description": "Forbidden",
                        "type": "string"
                    }
                },
                "type": "string"
            },
            "Notfound": {
                "title": "Not found",
                "description": "Not found",
                "properties": {
                    "notfound": {
                        "title": "Not found",
                        "description": "Not found",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CategoriesResponse": {
                "title": "Categories Response",
                "description": "Categories Response",
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "id": "3",
                            "name": "test test",
                            "slug": "test-test",
                            "status": "1",
                            "feature_image": "https://petparent.s3.ap-south-1.amazonaws.com/images/wRPTtNmKj0Kl8NFnUhuoLh5SyHpDXF5C7DmFzbS0.jpg",
                            "created": "Dec 10, 2021 11:34:05",
                            "updated_at": "2021-12-10T11:34:05.000000Z",
                            "childrens": "[]"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "Token": {
                "type": "apiKey",
                "name": "Token",
                "in": "header",
                "scheme": "Token"
            }
        }
    },
...
    "security": {
        "0": [],
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://127.0.0.1:8000//oauth/authorize",
                    "tokenUrl": "http://127.0.0.1:8000//oauth/token",
                    "refreshUrl": "http://127.0.0.1:8000//token/refresh",
                    "scopes": []
                }
            }
        }
    }
}